// src/app/Models/Shared/SupportingInfo.ts
export interface SupportingInfo {
    EmployeeMaxSize: number;
    EmployeeId: number;
    VideoCvContent: number;
    BannerShow: number;
    GoToNewJobPostings: boolean;
    GoToNewJobPostingsNewCompany: boolean;
    EntrepreneurCompany: boolean;
    EntrepreneurJobMail: boolean;
    RecruiterNewView: boolean;
    EmployerAppPromotion: boolean;
    EmployerApplicantProcessPromotion: boolean;
    ContactPerson: string;
    Name: string;
    Email: string;
    BillingContact: string;
    CompanyCountry: string;
    CreatedDate: string;
    UserName: string;
    Nd: boolean;
    ClosingLoopAccess: boolean;
    CompanyCity: string;
    AdminUserType: boolean;
    VerificationStatus: boolean;
    Feedback: boolean;
    NidVerify: boolean;
    SalesPersonName: string;
    SalesPersonDesignation: string;
    SalesPersonContact: string;
    SalesPersonEmail: string;
    SalesPersonImage: string;
    CvSearchAccess: boolean;
    JobPostingAccess: boolean;
    RemainingDaysCv: number | null;
    RemainingDaysJob: number | null;
    RemainingCv: string | null;
    JobPostingLimit: string | null;
    CvSearchService: {
      StartingDate: string | null;
      EndingDate: string | null;
      Id: number | null;
      Limit: number | null;
      Available: number | null;
      Viewed: number | null;
    };
    SmsPurchased: number;
    SmsSend: number;
    SmsRemaining: number;
    InvitedJobAccess: boolean;
    CompanyLiveJob: boolean;
    PaymentProcess: boolean;
    CompanySurvey: boolean;
    CompanySurveyID: string;
    CorporateAssessment: boolean;
    AssessmentPopupDisplay: boolean;
    Assessment2ndDisplay: string;
    CustomizeUrl: boolean;
    SmsPackage: boolean;
    OnlineTestService: boolean;
    UserAccessManagement: boolean;
    CorporateEmailService: boolean;
    AccessHighPrioritySignal: boolean;
    ApiAccessCvExport: boolean;
    CompanyActiveLogoURL: string;
    LastPostingDate: string;
    TotalPostedJob: number;
    TotalDraftedJob: number;
    TotalArchiveJob: number;
    TotalJobfairJob: number;
    TotalJobForUser: number;
    JobPostingService: {
      StartingDate: string | null;
      EndingDate: string | null;
      BasicListLimit: number | null;
      Id: number | null;
      MaxJob: number | null;
      MaxStandout: number | null;
      StandoutLimit: number | null;
      StandoutPremiumLimit: number | null;
      MaxStandoutPremium: number | null;
      BasicWhiteCollarLimit: number | null;
      StandOutCollarLimit: number | null;
      StandOutPremiumCollarLimit: number | null;
      BasicBlueCollarLimit: number | null;
      StandOutBlueCollarLimit: number | null;
      StandOutPremiumBlueCollarLimit: number | null;
      BasicBlueCollarPosted: number | null;
      StandOutBlueCollarPosted: number | null;
      StandOutPremiumBlueCollarPosted: number | null;
      JobPostingAccessForWhiteColler: number | null;
    };
  }
  